ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
insightface
onnxruntime
numpy
opencv-python
Pillow
geopy
python-dateutil
